check

+++ app/Http/Controllers/EmployeeController.php
@@ -4,10 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Employee;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class EmployeeController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index(){
+        if(!Auth::user()->can('index employee')){
+            return redirect()->back()->with([
+                'error' => 'Unauthorized'
+            ]);
+        }
         try{
             return view('employee.index')->with([
                 'title' => 'Employees',
@@ -24,6 +35,11 @@ class EmployeeController extends Controller
     }
 
     public function store(Request $request){
+        if(!Auth::user()->can('store employee')){
+            return redirect()->back()->with([
+                'error' => 'Unauthorized'
+            ]);
+        }
         try{
          Employee::create($request->except(['_token','_method']));
          return redirect()->back()->with([
@@ -37,6 +53,11 @@ class EmployeeController extends Controller
     }
 
     public function edit($id){
+        if(!Auth::user()->can('edit employee')){
+            return redirect()->back()->with([
+                'error' => 'Unauthorized'
+            ]);
+        }
         try{
             $emp = Employee::findOrFail($id);
             return view('employee.edit')->with([
@@ -54,6 +75,11 @@ class EmployeeController extends Controller
     }
 
     public function update($id, Request $request){
+        if(!Auth::user()->can('update employee')){
+            return redirect()->back()->with([
+                'error' => 'Unauthorized'
+            ]);
+        }
         try{
             Employee::findOrFail($id)->update($request->except(['_token','_method']));
             return redirect()->back()->with([
@@ -66,6 +92,11 @@ class EmployeeController extends Controller
         }
     }
     public function delete($id){
+        if(!Auth::user()->can('delete employee')){
+            return redirect()->back()->with([
+                'error' => 'Unauthorized'
+            ]);
+        }
         try{
             Employee::findOrFail($id)->delete();
             return redirect()->back()->with([
